fix(parser): parse version string properly
The version values in TOML have two valid forms:
  - simple string
anchor-lang = "0.18.2"
  - detailed table format
anchor-lang = { version = "0.18.2", features = ["derive"] }

The parser code used a hack (to the forked toml lib) when detecting the
above pattern, as it attempted to use a table form (i.e. `contains`) to
parse the value.

This change better handles the detection thus drops the hack.